requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision
PyQt6
transformers
Pillow
opencv-python
numpy
scikit-learn
tqdm
tensorrt
onnx
onnxruntime-gpu
onnxscript
send2trash
